refactor: answer/domain/dto 패키지를 answer/dto로 이동
