+++ src/main/java/com/upstage/devup/user/answer/service/UserAnswerService.java
@@ -1,6 +1,6 @@
 package com.upstage.devup.user.answer.service;
 
-import com.upstage.devup.answer.domain.dto.UserAnswerDetailDto;
+import com.upstage.devup.answer.dto.UserAnswerDetailDto;
 import com.upstage.devup.user.answer.domain.entity.UserAnswer;
 import com.upstage.devup.user.answer.domain.entity.UserWrongAnswer;
 import com.upstage.devup.user.answer.repository.UserAnswerRepository;
